# swagger/swagger.yaml
basePath: /
definitions:
...
  logtrace.Organization:
    properties:
      created_at:
        type: string
      id:
        type: string
      image_url:
        type: string
      is_active:
        type: boolean
      is_subscription_active:
        type: boolean
      name:
        type: string
      plan_name:
        type: string
      updated_at:
        type: string
    type: object
...
  logtrace.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_verified_at:
        type: string
      full_name:
        type: string
      id:
        type: string
      metadata:
        $ref: '#/definitions/logtrace.UserMetadata'
      phone:
        type: string
      roles:
        items:
          $ref: '#/definitions/logtrace.UserRole'
        type: array
      status:
        type: string
      updated_at:
        type: string
    type: object
  logtrace.UserMetadata:
    properties:
      organization_id:
        items:
          type: string
        type: array
      user_role:
        type: string
    type: object
  logtrace.UserRole:
    properties:
      assigned_at:
        type: string
      id:
        type: string
      role_name:
        type: string
      user_id:
        type: string
    type: object
  server.APIStatus:
    properties:
      message:
        type: string
    required:
    - message
    type: object
...
  server.fetchUserResponse:
    properties:
      message:
        type: string
      organization:
        $ref: '#/definitions/logtrace.Organization'
      organizations:
        items:
          $ref: '#/definitions/logtrace.Organization'
        type: array
      token:
        type: string
      user:
        $ref: '#/definitions/logtrace.User'
    required:
    - message
    type: object
  server.inviteUserRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      role:
        type: string
    type: object
...
  server.listUsersResponse:
    properties:
      message:
        type: string
      meta:
        $ref: '#/definitions/server.meta'
      users:
        items:
          $ref: '#/definitions/logtrace.User'
        type: array
    required:
    - message
    type: object
  server.loginRequest:
    properties:
      code:
        type: string
      email:
        type: string
      password:
        type: string
    type: object
  server.meta:
    properties:
      paging:
        $ref: '#/definitions/server.pagingInfo'
    required:
    - paging
    type: object
  server.pagingInfo:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    required:
    - page
    - per_page
    - total
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Logtrace API documentation
  title: Logtrace API
  version: 0.1.0
paths:
  /auth/account/{reference}:
    patch:
      consumes:
      - application/json
      description: Revoke a user's admin role in the organization.
      parameters:
      - description: User Reference (UUID)
        in: path
        name: reference
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User role revoked successfully
          schema:
            $ref: '#/definitions/server.APIStatus'
        "400":
          description: Invalid user reference
          schema:
            $ref: '#/definitions/server.APIStatus'
        "403":
          description: User does not belong to your organization
          schema:
            $ref: '#/definitions/server.APIStatus'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/server.APIStatus'
        "500":
          description: Could not update user role
          schema:
            $ref: '#/definitions/server.APIStatus'
      tags:
      - Users
  /auth/invite:
    post:
      consumes:
      - application/json
      description: Invite a user by email.
      parameters:
      - description: Invite User Request Body
        in: body
        name: inviteUserRequest
        required: true
        schema:
          $ref: '#/definitions/server.inviteUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Invitation sent successfully
          schema:
            $ref: '#/definitions/server.APIStatus'
        "400":
          description: invalid request body
          schema:
            $ref: '#/definitions/server.APIStatus'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/server.APIStatus'
      tags:
      - Users
  /auth/login/{provider}:
    post:
      consumes:
      - application/json
      description: Login handles user authentication via OAuth2 providers.
      parameters:
      - description: OAuth2 Provider (e.g., google)
        in: path
        name: provider
        required: true
        type: string
      - description: Login Request Body
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/server.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/server.fetchUserResponse'
        "400":
          description: Bad request or invalid credentials
          schema:
            $ref: '#/definitions/server.APIStatus'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.APIStatus'
      tags:
      - Authenticating
  /auth/me:
    get:
      consumes:
      - application/json
      description: FetchCurrentUser retrieves the currently authenticated user's details.
      produces:
      - application/json
      responses:
        "200":
          description: Fetched current user
          schema:
            $ref: '#/definitions/server.fetchUserResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.APIStatus'
      tags:
      - Users
  /auth/me/users:
    get:
      consumes:
      - application/json
      description: List all users in the organization.
      produces:
      - application/json
      responses:
        "200":
          description: List of organization users
          schema:
            $ref: '#/definitions/server.listUsersResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.APIStatus'
      tags:
      - Users
...
schemes:
- http
- https
swagger: "2.0"